fix(clerk-sdk-node): Fix circular import
