+++ packages/sdk-node/src/clerkClient.ts
@@ -3,28 +3,7 @@ import { Clerk as _Clerk, decodeJwt, verifyToken as _verifyToken } from '@clerk/
 
 import { createClerkExpressRequireAuth } from './clerkExpressRequireAuth';
 import { createClerkExpressWithAuth } from './clerkExpressWithAuth';
-
-export const loadClientEnv = () => {
-  return {
-    publishableKey: process.env.CLERK_PUBLISHABLE_KEY || '',
-    frontendApi: process.env.CLERK_FRONTEND_API || '',
-    clerkJSUrl: process.env.CLERK_JS || '',
-    clerkJSVersion: process.env.CLERK_JS_VERSION || '',
-  };
-};
-
-export const loadApiEnv = () => {
-  return {
-    secretKey: process.env.CLERK_SECRET_KEY || process.env.CLERK_API_KEY || '',
-    apiKey: process.env.CLERK_SECRET_KEY || process.env.CLERK_API_KEY || '',
-    apiUrl: process.env.CLERK_API_URL || 'https://api.clerk.dev',
-    apiVersion: process.env.CLERK_API_VERSION || 'v1',
-    domain: process.env.CLERK_DOMAIN || '',
-    proxyUrl: process.env.CLERK_PROXY_URL || '',
-    signInUrl: process.env.CLERK_SIGN_IN_URL || '',
-    isSatellite: process.env.CLERK_IS_SATELLITE === 'true',
-  };
-};
+import { loadApiEnv, loadClientEnv } from './utils';
 
 /**
  * This needs to be a *named* function in order to support the older
