+++ packages/sdk-node/tsconfig.json
@@ -1,12 +1,11 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "declarationDir": "dist/types",
-    "declarationMap": true,
-    "emitDeclarationOnly": true,
+    "declarationMap": false,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "importHelpers": true,
+    "isolatedModules": true,
     "moduleResolution": "node",
     "noImplicitReturns": true,
     "noUnusedLocals": true,
@@ -15,6 +14,7 @@
     "sourceMap": false,
     "strict": true,
     "target": "ES2020",
+    "outDir": "dist",
     "types": ["jest"]
   },
   "exclude": ["node_modules"],
